+++ t/apicast-path-routing.t
@@ -51,8 +51,6 @@ env APICAST_PATH_ROUTING_ENABLED;
 --- config
   include $TEST_NGINX_APICAST_CONFIG;
 
-  set $backend_endpoint 'http://127.0.0.1:$TEST_NGINX_SERVER_PORT';
-
   location /transactions/authrep.xml {
     content_by_lua_block { ngx.exit(200) }
   }
@@ -127,8 +125,6 @@ env APICAST_PATH_ROUTING_ENABLED;
   lua_shared_dict api_keys 10m;
 --- config
   include $TEST_NGINX_APICAST_CONFIG;
-
-  set $backend_endpoint 'http://127.0.0.1:$TEST_NGINX_SERVER_PORT';
 
   location /transactions/authrep.xml {
     content_by_lua_block { ngx.exit(200) }